refactor(domain): move dendrogram logic
