+++ src/components/Dendrogram/State.ts
@@ -1,0 +1,4 @@
+import * as TreeLink from "../TreeLink";
+import * as TreeNode from "../TreeNode";
+
+export type State = TreeLink.State & TreeNode.State;
