+++ src/components/Menu/Store.ts
@@ -6,20 +6,13 @@ export interface Store {
   updateKey: (key: string) => void;
 }
 
-export const generateStore = (domainStores: Domain.Stores): Store => {
-  return {
-    currentKey: domainStores.app.state.currentKey,
-    scripts: domainStores.app.state.scripts,
-    updateKey: (key: string) => {
-      domainStores.app.dispatch({
-        type: "UPDATE_KEY",
-        currentKey: key,
-      });
-      domainStores.dendrogram.dispatch({
-        type: "UPDATE_KEY",
-        currentKey: key,
-        pkg: domainStores.app.state.pkg,
-      });
-    },
-  };
-};
+export const generateStore = (domainStores: Domain.Stores): Store => ({
+  currentKey: domainStores.app.state.currentKey,
+  scripts: domainStores.app.state.scripts,
+  updateKey: (key: string) => {
+    domainStores.app.dispatch({
+      type: "UPDATE_KEY",
+      currentKey: key,
+    });
+  },
+});
