+++ src/components/TreeLink/State.ts
@@ -2,8 +2,5 @@ import * as Types from "@this/types";
 
 export interface State {
   scale: Types.Adjustment;
-  nodes: Types.Node[];
   links: Types.Link[];
-  radius: number;
-  offset: number;
 }
