+++ src/components/TreeLink/index.tsx
@@ -1,3 +1,4 @@
 export { Container } from "./Container";
 export { generateStore, Store } from "./Store";
+export * from "./State";
 export * from "./TreeLink";
