+++ src/components/TreeNode/State.ts
@@ -1,0 +1,8 @@
+import * as Types from "@this/types";
+
+export interface State {
+  scale: Types.Adjustment;
+  nodes: Types.Node[];
+  radius: number;
+  offset: number;
+}
