+++ src/components/TreeNode/Store.ts
@@ -1,5 +1,6 @@
 import * as Domain from "@domain";
 import * as Types from "@this/types";
+import { State } from "./State";
 
 export interface Store {
   updateKey: (key: string) => void;
@@ -9,22 +10,15 @@ export interface Store {
   nodes: Types.Node[];
 }
 
-export const generateStore = (domainStores: Domain.Stores): Store => {
-  return {
-    updateKey: (key: string) => {
-      domainStores.app.dispatch({
-        type: "UPDATE_KEY",
-        currentKey: key,
-      });
-      domainStores.dendrogram.dispatch({
-        type: "UPDATE_KEY",
-        currentKey: key,
-        pkg: domainStores.app.state.pkg,
-      });
-    },
-    scale: domainStores.dendrogram.state.scale,
-    radius: domainStores.dendrogram.state.radius,
-    offset: domainStores.dendrogram.state.offset,
-    nodes: domainStores.dendrogram.state.nodes,
-  };
-};
+export const generateStore = (domainStores: Domain.Stores, state: State): Store => ({
+  updateKey: (key: string) => {
+    domainStores.app.dispatch({
+      type: "UPDATE_KEY",
+      currentKey: key,
+    });
+  },
+  scale: state.scale,
+  radius: state.radius,
+  offset: state.offset,
+  nodes: state.nodes,
+});
