+++ src/components/TreeNode/index.ts
@@ -1,3 +1,4 @@
 export { Container } from "./Container";
 export { generateStore, Store } from "./Store";
+export * from "./State";
 export * from "./TreeNode";
