--- src/domain/Dendrogram/index.ts
@@ -1,3 +1,0 @@
-export * from "./State";
-export * from "./Reducer";
-export * from "./Action";
